Move `LiveHackrNewsViewModel` to its own file
