+++ HackrNewsiOS/Stories UI/Controllers/LiveHackrNewsRefreshController.swift
@@ -5,60 +5,13 @@
 import HackrNews
 import UIKit
 
-final class LiveHackrNewsModel {
-    private let loader: LiveHackrNewsLoader
-
-    init(loader: LiveHackrNewsLoader) {
-        self.loader = loader
-    }
-
-    private enum State {
-        case pending
-        case loading
-        case loaded([LiveHackrNew])
-        case failed
-    }
-
-    private var state: State = .pending {
-        didSet {
-            onChange?(self)
-        }
-    }
-
-    var onChange: ((LiveHackrNewsModel) -> Void)?
-
-    var isLoading: Bool {
-        switch state {
-        case .loading: return true
-        case .pending, .loaded, .failed: return false
-        }
-    }
-
-    var news: [LiveHackrNew]? {
-        switch state {
-        case let .loaded(news): return news
-        case .pending, .loading, .failed: return nil
-        }
-    }
-
-    func loadNews() {
-        loader.load { [weak self] result in
-            if let news = try? result.get() {
-                self?.state = .loaded(news)
-            } else {
-                self?.state = .failed
-            }
-        }
-    }
-}
-
 final class LiveHackrNewsRefreshController: NSObject {
     private(set) lazy var view = binded(UIRefreshControl())
 
-    private let viewModel: LiveHackrNewsModel
+    private let viewModel: LiveHackrNewsViewModel
 
     init(loader: LiveHackrNewsLoader) {
-        viewModel = LiveHackrNewsModel(loader: loader)
+        viewModel = LiveHackrNewsViewModel(loader: loader)
     }
 
     var onRefresh: (([LiveHackrNew]) -> Void)?
